changed code for new snb website
