+++ exchange_rate_scraper/__init__.py
@@ -1,0 +1,4 @@
+import sys
+import os
+import exchange_rate_scraper.constants
+import exchange_rate_scraper.main
