+++ pyproject.toml
@@ -1,0 +1,73 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "my_package"
+authors = [
+    {name = "Luca Mircea", email = "lucamircea10@gmail.com"},
+]
+description = "Python script for scraping exchange rates from central bank websites"
+readme = "README.md"
+requires-python = ">=3.12"
+keywords = ["one", "two"]
+license = {text = "BSD-3-Clause"}
+classifiers = [
+    "Programming Language :: Python :: 3",
+]
+
+dependencies = [
+    "attrs==23.1.0",
+    "beautifulsoup4==4.12.2",
+    "black==23.9.1",
+    "boto3==1.28.57",
+    "botocore==1.31.57",
+    "certifi==2023.7.22",
+    "cffi==1.15.1",
+    "charset-normalizer==3.2.0",
+    "click==8.1.7",
+    "colorama==0.4.6",
+    "cssselect==1.2.0",
+    "exceptiongroup==1.1.3",
+    "h11==0.14.0",
+    "idna==3.4",
+    "isort==5.12.0",
+    "jmespath==1.0.1",
+    "lxml==4.9.3",
+    "mypy-extensions==1.0.0",
+    "numpy==1.26.0",
+    "outcome==1.2.0",
+    "packaging==23.1",
+    "pandas==2.1.1",
+    "parsel==1.8.1",
+    "pathspec==0.11.2",
+    "platformdirs==3.10.0",
+    "pycparser==2.21",
+    "PySocks==1.7.1",
+    "python-dateutil==2.8.2",
+    "python-dotenv==1.0.0",
+    "pytz==2023.3.post1",
+    "requests==2.31.0",
+    "ruff==0.0.291",
+    "s3transfer==0.7.0",
+    "selenium==4.12.0",
+    "six==1.16.0",
+    "sniffio==1.3.0",
+    "sortedcontainers==2.4.0",
+    "soupsieve==2.5",
+    "trio==0.22.2",
+    "trio-websocket==0.10.4",
+    "tzdata==2023.3",
+    "urllib3==1.26.16",
+    "w3lib==2.1.2",
+    "webdriver-manager==4.0.1",
+    "wsproto==1.2.0"
+]
+dynamic = ["version"]
+
+[project.optional-dependencies]
+pdf = ["ReportLab>=1.2", "RXP"]
+rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]
+
+[project.scripts]
+forex-scraper = "exchange_rate_scraper.main:main"
